바운더리 박스 탭 추가

+++ src/main/java/io/aetherit/kfashion/ws/model/KfashionRectList.java
@@ -12,12 +12,12 @@ import lombok.NoArgsConstructor;
 public class KfashionRectList {
 
     private int id;
-    private int left;
-    private int top;
+    private float left;
+    private float top;
     private int width;
     private int height;
-    private Double scaleX;
-    private Double scaleY;
+    private float scaleX;
+    private float scaleY;
 
     private String createdId;
     private Long workNo;
